Match Structures

Match Structures

match_structures(
  query_compounds,
  target_compound_structures,
  target_compound_aliases = NULL,
  match_var_order = c("inchikey", "inchikey_noproto", "inchikey_connectivity", "name"),
  target_primary_key = "systematicCompoundId",
  drop_ambiguous = TRUE
)

Arguments

query_compounds

a tibble of compounds containing some combination of: name, inchi, inchikey, smiles.

target_compound_structures

a tibble of compounds containing a systematicCompoundId and some combination of: name, inchi, inchikey, smiles.

target_compound_aliases

an optional tibble containing additional aliases for a compound (1-many) with fields systematicCompoundId and alias.

match_var_order

a character vector of identifiers that will be sequentially searched for matches to a query_compound until a match is found.

Options are

inchikey: exact match the full inchikey

inchikey_noproto: inchikey without the protonation layer (this field will be generated from the inchikey)

inchikey_connectivity: exact match to the structure layer of the inchikey (this field will be generated from the inchikey)

inchi: exact match to inchi

smiles: exact match to smiles

name: exact match to a compound's name

alias: exact match to a compound's aliases (from target_compound_aliases)

target_primary_key

variable in target_compound_structures and target_compound_aliases if provided which defines a unique compound.

drop_ambiguous

if drop_ambiguous is TRUE then an ambiguous match of one query to 2+ distinct compounds in systematic_compounds_con for a given variable in match_var_order will not be considered a match. If FALSE then multiple database matches may be returned for a given query.

Value

query_compounds with a systematicCompoundId column indicating which entry of target_compounds matches.