Query Systematic Compounds

Query Systematic Compounds

query_systematic_compounds(
  systematic_compounds_con,
  systematicCompoundIds = NULL,
  db_update_version = "recent"
)

Arguments

systematic_compounds_con

A connection to an SQL database which stores the structure and systematic IDs of well-studied compounds.

systematicCompoundIds

NULL to query all compounds or a numeric vector of systematic compound IDs

db_update_version

multiple versions of the database exist that can be differentiated by update date.

all

Queries all versions of the database. This is useful when previously-created systematicCompoundIds need to be matched

recent

Queries the most recent version of the database. This is useful when matching to compounds by name or structure.

YYYY-MM-DD

Provide a specific date with the format YYYY-MM-DD to query database entries from this date.

Value

a list containing:

distinct_compounds

unique database entries

database_ids

systematic IDs of compounds in KEGG, HMDB, ...

compound_aliases

alternative names for a compound

metabolic_pathways

KEGG pathways associated with a metabolite