Populate Standard DB Data

Add new mass spectrometry standard data present in ms_data_list to the mass_spec_standards_con database respecting existing entries and looking up unique compounds in ms_data_list.

standard_db_populate_data(
  ms_data_list,
  mass_spec_standards_con,
  systematic_compounds_con,
  annotationSource,
  match_var_order = c("inchikey", "name", "alias", "inchikey_noproto",
    "inchikey_connectivity"),
  write_to_db = FALSE,
  strict = TRUE
)

Arguments

ms_data_list

a list containg mass spectrometry experimental data. Three entires are required: ions, samples and measurements:

ions

samples

elutions [optional]

fragmentations [optional]

fragmentationData [optional]

mass_spec_standards_con

A connection to an SQL database which stores experimental mass spectrometry data.

systematic_compounds_con

A connection to an SQL databasing housing names, structure, aliases and systematic identifiers of known compounds.

annotationSource

A unique character string to label the data source.

match_var_order

identifier priority order for matching.

write_to_db

if FALSE then return the entries which would be appended to the database, if TRUE then append these to mass_spec_standards_con.

strict

return warnings in some cases rather than errors.

Value

a list of tables (if write_to_db is FALSE) or if write_to_db is TRUE: append new table entries to mass_spec_standards_con as side-effects and return 1.

Details

All supplied tables must include a .entry_id identifier which indicates the relationship between tables in the original dataset. This function will build mappings between .entry_id and the correponding sampleId, compoundId, ionId, elutionId and fragmentationId and append new entries to these tables as necessary.

All samples and measurements will be added but only ions and compounds which are not already found in the database.

See also