Summarize Compounds

Reduce a set of compounds associated with data to unique entries by name, structures, and database aliases.

standard_db_summarize_compounds(
  ms_data_ions,
  mass_spec_standards_con,
  systematic_compounds_con,
  match_var_order = c("inchikey", "name", "alias", "inchikey_noproto",
    "inchikey_connectivity"),
  annotationSource
)

Arguments

ms_data_ions

ions table from ms_data_list.

mass_spec_standards_con

A connection to an SQL database which stores experimental mass spectrometry data.

systematic_compounds_con

A connection to an SQL databasing housing names, structure, aliases and systematic identifiers of known compounds.

match_var_order

identifier priority order for matching.

annotationSource

A unique character string to label the data source.

Value

entry_compound_dict

entry_id to compoundId

compound_attr

compoundId attributes - e.g., systematic id, structure, name, formula

compound_annotation_sources

compound_annotation_sources - annotations associated with a compoundId for a given annotationSource

compoundId_conversions

rewrite databased compoundIds (if they fused with another w/ further annotation)

standard_inconsistencies

multiple compoundIds fusing based on annotations

systematic_inconsistencies

multiple systematc compounds matching to a single compoundId

Details

Compounds are split into two categories - those with identifying fields (name and/or structure) and those without any identifying fields. ions without identifying fields are each stored as a unique entry. ions with identifying fields are grouped by name + structure and compared to standards and systematic compounds.